+++ packages/sqle/src/testUtils/mockRequest.ts
@@ -1,11 +1,7 @@
 import { AxiosResponse } from 'axios';
-import user from '@actiontech/shared/lib/api/sqle/service/user';
-import instance from '@actiontech/shared/lib/api/sqle/service/instance';
 import ruleTemplate from '@actiontech/shared/lib/api/sqle/service/rule_template';
 import configuration from '@actiontech/shared/lib/api/sqle/service/configuration';
-import operation from '@actiontech/shared/lib/api/sqle/service/operation';
 import audit_plan from '@actiontech/shared/lib/api/sqle/service/audit_plan';
-import OperationRecord from '@actiontech/shared/lib/api/sqle/service/OperationRecord';
 import { driverMeta } from '../hooks/useDatabaseType/index.test.data';
 
 export const successData = (data: any, otherData?: any) => {
@@ -115,188 +111,9 @@ export const rejectThreeSecond = (
   });
 };
 
-export const throwErrorThreeSecond = (
-  error = 'error',
-  { status = 500, headers = {}, config = {}, statusText = '' } = {}
-) => {
-  return new Promise((res) => {
-    throw new Error(error);
-  });
-};
-
-export const mockUseUsername = () => {
-  const spy = jest.spyOn(user, 'getUserTipListV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([{ user_name: 'user_name1' }])
-  );
-  return spy;
-};
-
-export const mockUseInstance = () => {
-  const spy = jest.spyOn(instance, 'getInstanceTipListV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([{ instance_name: 'instance1', instance_type: 'mysql' }])
-  );
-  return spy;
-};
-
-export const mockUseInstanceSchema = () => {
-  const spy = jest.spyOn(instance, 'getInstanceSchemasV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond({ schema_name_list: ['schema1'] })
-  );
-  return spy;
-};
-
-// export const mockUseRole = () => {
-//   const spy = jest.spyOn(role, 'getRoleTipListV1');
-//   spy.mockImplementation(() =>
-//     resolveThreeSecond([{ role_name: 'role_name1' }])
-//   );
-//   return spy;
-// };
-
-export const mockUseOperation = () => {
-  const spy = jest.spyOn(operation, 'GetOperationsV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      {
-        op_code: 20100,
-        op_desc: '查看工单'
-      },
-      {
-        op_code: 20150,
-        op_desc: '查看他人创建的工单'
-      },
-      {
-        op_code: 20200,
-        op_desc: '更新工单'
-      },
-      {
-        op_code: 20300,
-        op_desc: '创建工单'
-      },
-      {
-        op_code: 20400,
-        op_desc: '删除工单'
-      }
-    ])
-  );
-  return spy;
-};
-
-export const mockUseRuleTemplate = () => {
-  const spy = jest.spyOn(ruleTemplate, 'getProjectRuleTemplateTipsV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      {
-        rule_template_name: 'rule_template_name1',
-        rule_template_id: '1',
-        db_type: 'mysql'
-      },
-      {
-        rule_template_name: 'rule_template_name2',
-        rule_template_id: '2',
-        db_type: 'oracle'
-      }
-    ])
-  );
-  return spy;
-};
-
-export const mockUseGlobalRuleTemplate = () => {
-  const spy = jest.spyOn(ruleTemplate, 'getRuleTemplateTipsV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      {
-        rule_template_name: 'global_rule_template_name1',
-        rule_template_id: '01',
-        db_type: 'mysql'
-      },
-      {
-        rule_template_name: 'global_rule_template_name2',
-        rule_template_id: '02',
-        db_type: 'oracle'
-      }
-    ])
-  );
-  return spy;
-};
-
 export const mockDriver = () => {
   const spy = jest.spyOn(configuration, 'getDriversV2');
   spy.mockImplementation(() => resolveThreeSecond(driverMeta));
-  return spy;
-};
-
-export const mockInstanceTip = () => {
-  const spy = jest.spyOn(instance, 'getInstanceTipListV1');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      {
-        instance_name: 'mysql-test',
-        instance_type: 'mysql'
-      },
-      {
-        instance_name: 'oracle-test',
-        instance_type: 'oracle'
-      }
-    ])
-  );
-  return spy;
-};
-
-// export const mockManagerPermission = () => {
-//   const spy = jest.spyOn(management_permission, 'GetManagementPermissionsV1');
-//   spy.mockImplementation(() =>
-//     resolveThreeSecond([
-//       {
-//         code: 1,
-//         desc: '创建项目'
-//       }
-//     ])
-//   );
-//   return spy;
-// };
-
-// export const mockUseTaskSource = () => {
-//   const spy = jest.spyOn(sync_instance, 'GetSyncTaskSourceTips');
-//   spy.mockImplementation(() =>
-//     resolveThreeSecond([
-//       { source: 'source1', db_types: ['mysql'] },
-//       { source: 'source2', db_types: ['oracle'] }
-//     ])
-//   );
-//   return spy;
-// };
-
-export const mockUseOperationTypeName = () => {
-  const spy = jest.spyOn(OperationRecord, 'GetOperationTypeNameList');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      { operation_type_name: 'project', desc: '项目' },
-      { operation_type_name: 'instance', desc: '数据源' }
-    ])
-  );
-  return spy;
-};
-
-export const mockUseOperationActions = () => {
-  const spy = jest.spyOn(OperationRecord, 'getOperationActionList');
-  spy.mockImplementation(() =>
-    resolveThreeSecond([
-      {
-        operation_action: 'edit_instance',
-        desc: '编辑数据源',
-        operation_type: 'instance'
-      },
-      {
-        operation_action: 'create_project',
-        desc: '创建项目',
-        operation_type: 'project'
-      }
-    ])
-  );
   return spy;
 };
