+++ packages/sqle/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src"],
+  "include": ["src", "../shared/lib/testUtil/mockAntDesignPlots.jsx", "../shared/lib/testUtil/mockEditor.jsx"],
   "compilerOptions": {
     "paths": {
       "~/*": ["./src/*"]
